add flush-to-backend step before close-notify
When a client sends a DTLS packet and then immediately closes the DTLS
session, there is a race condition in which the session will be freed
before the packet is sent to the server.

We added a step in which as long as there are packets to send to the
server (usually 0 or 1) they will be sent, and only then we proceed to
the close-notify step.